propagated changes of Thursday Mar 01 2007 - Friday Mar 02 2007 from the MonetDB_1-16 branch to the development trunk

+++ buildtools/Mx/Display.c
@@ -86,7 +86,7 @@ PrModeStr(char *s, char c)
 void
 PrCmd(char *s)
 {
-	extern int pr_pos;
+	extern size_t pr_pos;
 
 	if (Hide())
 		return;
